requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-openai==0.2.0
langchain-google-genai>=2.0.0
langsmith<0.2.0
numpy<2.0.0
huggingface-hub<1.0.0
unstructured[all-docs]
pdf2image
pytesseract
pillow
lxml
tiktoken
chromadb
streamlit
pydantic
python-dotenv